requirements.txt
numpy~=2.0.2
future~=0.18.2
pillow~=11.1.0
setuptools~=58.0.4
requests~=2.32.5